fix(*): implemented lazy loading on dashboard
